+++ hat/examples/violajones/pom.xml
@@ -16,11 +16,11 @@ version 2 for more details (a copy is included in the LICENSE file that
 
 You should have received a copy of the GNU General Public License version
 2 along with this work; if not, write to the Free Software Foundation,
-  Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
 
-     Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
-     or visit www.oracle.com if you need additional information or have any
-     questions.
+Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+or visit www.oracle.com if you need additional information or have any
+questions.
 --><project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0">
     <!--Auto generated by mkpoms-->
     <modelVersion>4.0.0</modelVersion>
