+++ hat/mkpoms
@@ -25,9 +25,7 @@
  */
 
 import static bldr.Bldr.*;
-import bldr.CMakeProbe;
-import bldr.Capabilities;
-import static bldr.Capabilities.*;
+import static bldr.Bldr.Capabilities.*;
 
 void main(String[] argv) {
    var pomComment = """
@@ -48,12 +46,12 @@ void main(String[] argv) {
 
       You should have received a copy of the GNU General Public License version
       2 along with this work; if not, write to the Free Software Foundation,
-        Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+      Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
 
-           Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
-           or visit www.oracle.com if you need additional information or have any
-           questions.                
-           """;
+      Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+      or visit www.oracle.com if you need additional information or have any
+      questions.                
+      """;
 
    var hatDir = Dir.current();
